requirements.txt
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.3.2
python-dotenv==0.7.1
python-socketio==1.8.1
